+++ HISTORY.md
@@ -4,5 +4,5 @@
 ## 1.4.0 / 2015-01-12
 
 - fix first option. 
-- add fieldFirst option.
+- add firstFields option.
 - see tests/validator.spec.js